requirements.txt
flask==3.0.3
line-bot-sdk==3.19.1
openai==1.109.1
httpx==0.27.2
gunicorn==21.2.0
filelock==3.16.1
